server/requirements.txt
facenet_pytorch==2.5.2
mmcv==1.3.14
ipython==7.27.0
notebook==6.4.4
onnx==1.10.1
mxnet==1.8.0.post0
numpy==1.21.2
matplotlib==3.4.3
opencv-python==4.5.3.56
scikit-learn==1.0
easydict==1.9
scikit-image==0.18.3
Flask==2.0.1
Flask_RESTful==0.3.9
faiss-gpu==1.7.1
gluoncv==0.10.4
mxnet==1.8.0
Flask_Cors==3.0.10